remove extended-instrumentations, use patching instead
