+++ aws-distro-opentelemetry-node-autoinstrumentation/src/register.ts
@@ -17,8 +17,8 @@ import { diag, DiagConsoleLogger, trace } from '@opentelemetry/api';
 import { getNodeAutoInstrumentations, InstrumentationConfigMap } from '@opentelemetry/auto-instrumentations-node';
 import { Instrumentation } from '@opentelemetry/instrumentation';
 import * as opentelemetry from '@opentelemetry/sdk-node';
-import { AwsOpentelemetryConfigurator, isLambdaEnvironment } from './aws-opentelemetry-configurator';
-import { applyInstrumentationPatches } from './patches/instrumentation-patch';
+import { AwsOpentelemetryConfigurator } from './aws-opentelemetry-configurator';
+import { applyInstrumentationPatches, customExtractor } from './patches/instrumentation-patch';
 
 diag.setLogger(new DiagConsoleLogger(), opentelemetry.core.getEnv().OTEL_LOG_LEVEL);
 
@@ -36,8 +36,6 @@ This file may also be used to apply patches to upstream instrumentation - usuall
 long-term changes to upstream.
 */
 
-let usePatchedAwsLambdaInstrumentation = false;
-
 export function setAwsDefaultEnvironmentVariables(): void {
   if (!process.env.OTEL_EXPORTER_OTLP_PROTOCOL) {
     process.env.OTEL_EXPORTER_OTLP_PROTOCOL = 'http/protobuf';
@@ -51,18 +49,13 @@ export function setAwsDefaultEnvironmentVariables(): void {
   if (!process.env.OTEL_NODE_DISABLED_INSTRUMENTATIONS) {
     process.env.OTEL_NODE_DISABLED_INSTRUMENTATIONS = 'fs,dns';
   }
-
-  // Temporary solution to never initialize the unpatched AWS Lambda Instrumentation,
-  // and only initialize the patched one in `src/patches/extended-instrumentations/aws-lambda.ts`
-  // This workaround is for until the patched Lambda Instrumentation is removed
-  if (!process.env.OTEL_NODE_DISABLED_INSTRUMENTATIONS.includes('aws-lambda') && isLambdaEnvironment()) {
-    process.env.OTEL_NODE_DISABLED_INSTRUMENTATIONS += ',aws-lambda';
-    usePatchedAwsLambdaInstrumentation = true;
-  }
 }
 setAwsDefaultEnvironmentVariables();
 
-const instrumentationConfigs: InstrumentationConfigMap = {
+export const instrumentationConfigs: InstrumentationConfigMap = {
+  '@opentelemetry/instrumentation-aws-lambda': {
+    eventContextExtractor: customExtractor,
+  },
   '@opentelemetry/instrumentation-aws-sdk': {
     suppressInternalInstrumentation: true,
   },
@@ -73,7 +66,7 @@ const instrumentationConfigs: InstrumentationConfigMap = {
 const instrumentations: Instrumentation[] = getNodeAutoInstrumentations(instrumentationConfigs);
 
 // Apply instrumentation patches
-applyInstrumentationPatches(instrumentations, instrumentationConfigs, usePatchedAwsLambdaInstrumentation);
+applyInstrumentationPatches(instrumentations);
 
 const configurator: AwsOpentelemetryConfigurator = new AwsOpentelemetryConfigurator(instrumentations, useXraySampler);
 const configuration: Partial<opentelemetry.NodeSDKConfiguration> = configurator.configure();
