+++ aws-distro-opentelemetry-node-autoinstrumentation/test/patches/instrumentation-patch.test.ts
@@ -10,6 +10,7 @@ import {
   Span,
   Tracer,
   AttributeValue,
+  TextMapSetter,
 } from '@opentelemetry/api';
 import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node';
 import { Instrumentation } from '@opentelemetry/instrumentation';
@@ -27,7 +28,7 @@ import { S3 } from '@aws-sdk/client-s3';
 import nock = require('nock');
 import { ReadableSpan } from '@opentelemetry/sdk-trace-base';
 import { getTestSpans } from '@opentelemetry/contrib-test-utils';
-import { AwsSdkInstrumentationExtended } from '../../src/patches/extended-instrumentations/aws-sdk-instrumentation-extended';
+import { instrumentationConfigs } from '../../src/register';
 
 // It is assumed that bedrock.test.ts has already registered the
 // necessary instrumentations for testing by calling:
@@ -55,12 +56,10 @@ const mockHeaders = {
   'content-type': 'application/json',
 };
 
-const UNPATCHED_INSTRUMENTATIONS: Instrumentation[] = getNodeAutoInstrumentations();
+const UNPATCHED_INSTRUMENTATIONS: Instrumentation[] = getNodeAutoInstrumentations(instrumentationConfigs);
 
-process.env.OTEL_NODE_DISABLED_INSTRUMENTATIONS += ',aws-lambda';
-const usePatchedAwsLambdaInstrumentation = true;
-const PATCHED_INSTRUMENTATIONS: Instrumentation[] = getNodeAutoInstrumentations();
-applyInstrumentationPatches(PATCHED_INSTRUMENTATIONS, {}, usePatchedAwsLambdaInstrumentation);
+const PATCHED_INSTRUMENTATIONS: Instrumentation[] = getNodeAutoInstrumentations(instrumentationConfigs);
+applyInstrumentationPatches(PATCHED_INSTRUMENTATIONS);
 
 describe('InstrumentationPatchTest', () => {
   it('SanityTestUnpatchedAwsSdkInstrumentation', () => {
@@ -118,9 +117,6 @@ describe('InstrumentationPatchTest', () => {
     expect(services.get('BedrockRuntime')).toBeTruthy();
     // Sanity check
     expect(services.has('InvalidService')).toBeFalsy();
-
-    // Check that the original AWS SDK Instrumentation is replaced with the extended version
-    expect(awsSdkInstrumentation).toBeInstanceOf(AwsSdkInstrumentationExtended);
   });
 
   it('SQS without patching', () => {
@@ -538,6 +534,36 @@ describe('InstrumentationPatchTest', () => {
   describe('AwsSdkInstrumentationPatchTest', () => {
     let s3: S3;
     const region = 'us-east-1';
+
+    it('overridden _getV3SmithyClientSendPatch updates MiddlewareStack', async () => {
+      const mockedMiddlewareStackInternal: any = [];
+      const mockedMiddlewareStack = {
+        add: (arg1: any, arg2: any) => mockedMiddlewareStackInternal.push([arg1, arg2]),
+      };
+      const send = extractAwsSdkInstrumentation(PATCHED_INSTRUMENTATIONS)
+        ['_getV3SmithyClientSendPatch']((...args: unknown[]) => Promise.resolve())
+        .bind({ middlewareStack: mockedMiddlewareStack });
+      sinon
+        .stub(AWSXRayPropagator.prototype, 'inject')
+        .callsFake((context: OtelContext, carrier: unknown, setter: TextMapSetter) => {
+          (carrier as any)['isCarrierModified'] = 'carrierIsModified';
+        });
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      await send({}, null);
+
+      const middlewareArgs: any = {
+        request: {
+          headers: {},
+        },
+      };
+      await mockedMiddlewareStackInternal[0][0]((arg: any) => Promise.resolve(), null)(middlewareArgs);
+
+      sinon.restore();
+      expect(middlewareArgs.request.headers['isCarrierModified']).toEqual('carrierIsModified');
+      expect(mockedMiddlewareStackInternal[0][1].name).toEqual('_adotInjectXrayContextMiddleware');
+    });
 
     it('injects trace context header into request via propagator', async () => {
       s3 = new S3({
